+++ examples/nuxt/server/middleware/federation.ts
@@ -1,0 +1,4 @@
+import { fedifyMiddleware } from "@fedify/nuxt";
+import federation from "../lib/federation";
+
+export default fedifyMiddleware(federation, (_event) => undefined);